+++ app/Http/Controllers/Shipments/ShipmentController.php
@@ -74,7 +74,6 @@ class ShipmentController extends ResourceSearchController
         return Inertia::render('Shipments/Show', [
             'shipment' => $shipment->load('carrier', 'shippers'),
             'stops' => ShipmentStopResource::collection($shipment->stops->load('facility')),
-            'notes' => Inertia::defer(fn () => NoteResource::collection($shipment->notes->load('user'))),
         ]);
     }
